feat: merrytoday check & delete chat on unmerry

+++ server/apps/merryList.js
@@ -106,6 +106,42 @@ merryListRouter.delete("/:user_id/delete", async (req, res) => {
     }
     const deleteId = data[0].id;
     await supabase.from("like_user").delete().eq("id", deleteId);
+
+    // const conversationId = await supabase
+    //   .from("conversation")
+    //   .select("id")
+    //   .eq("sender_id", user_id)
+    //   .eq("receiver_id", receivedIds);
+    // const messagesByConversationId = await supabase
+    //   .from("messages")
+    //   .select("id")
+    //   .eq("conversation_id", conversationId);
+    // await supabase
+    //   .from("conversation")
+    //   .delete()
+    //   .eq("sender_id", user_id)
+    //   .eq("receiver_id", receivedIds);
+    // await supabase
+    //   .from("messages")
+    //   .delete()
+    //   .eq("conversation_id", messagesByConversationId);
+
+    const conversationIdData = await supabase
+      .from("conversation")
+      .select("id")
+      .eq("sender_id", user_id)
+      .eq("receiver_id", receivedIds);
+    const conversationId = conversationIdData.data[0].id;
+    await supabase
+      .from("messages")
+      .delete()
+      .eq("conversation_id", conversationId);
+    await supabase
+      .from("conversation")
+      .delete()
+      .eq("sender_id", user_id)
+      .eq("receiver_id", receivedIds);
+
     return res.json({ message: "Unmerry Successfully" });
   } catch (error) {
     return res.status(500).json({ message: error.message });
